refactor: extract workspace subscribing
