+++ apps/keck/src/server/mod.rs
@@ -2,19 +2,20 @@ mod api;
 mod files;
 mod sync;
 mod utils;
+mod subscribe;
 
 use axum::{http::Method, Extension, Router, Server};
 use std::collections::HashMap;
 use std::thread::sleep;
-use std::time::Duration;
-use std::{net::SocketAddr, sync::Arc, thread};
+use std::{net::SocketAddr, sync::Arc};
 use tokio::runtime::Runtime;
 use tokio::sync::RwLock;
 use tokio::{runtime, signal};
 use tower_http::cors::{Any, CorsLayer};
 
 use api::Context;
-use utils::*;
+pub use utils::*;
+pub use subscribe::*;
 
 async fn shutdown_signal() {
     let ctrl_c = async {
@@ -63,22 +64,6 @@ impl WorkspaceChangedBlocks {
     }
 }
 
-fn generate_ws_callback(
-    workspace_changed_blocks: Arc<RwLock<HashMap<String, WorkspaceChangedBlocks>>>,
-    runtime: Arc<Runtime>,
-) -> Box<dyn Fn(String, Vec<String>) + Send + Sync> {
-    Box::new(move |workspace_id, block_ids| {
-        let workspace_changed_blocks = workspace_changed_blocks.clone();
-        runtime.spawn(async move {
-            let mut write_guard = workspace_changed_blocks.write().await;
-            write_guard
-                .entry(workspace_id.clone())
-                .or_insert(WorkspaceChangedBlocks::new(workspace_id.clone()))
-                .insert_block_ids(block_ids.clone());
-        });
-    })
-}
-
 pub async fn start_server() {
     let origins = [
         "http://localhost:4200".parse().unwrap(),
@@ -114,55 +99,12 @@ pub async fn start_server() {
     let workspace_changed_blocks =
         Arc::new(RwLock::new(HashMap::<String, WorkspaceChangedBlocks>::new()));
     let hook_endpoint = Arc::new(RwLock::new(String::new()));
-    {
-        let runtime = runtime.clone();
-        let workspace_changed_blocks = workspace_changed_blocks.clone();
-        let hook_endpoint = hook_endpoint.clone();
-        let client = client.clone();
-        thread::spawn(move || {
-            let runtime_cloned = runtime.clone();
-            loop {
-                let workspace_changed_blocks = workspace_changed_blocks.clone();
-                let hook_endpoint = hook_endpoint.clone();
-                let runtime_cloned = runtime_cloned.clone();
-                let client = client.clone();
-                runtime.spawn(async move {
-                    let read_guard = workspace_changed_blocks.read().await;
-                    if !read_guard.is_empty() {
-                        let endpoint = hook_endpoint.read().await;
-                        if !endpoint.is_empty() {
-                            {
-                                let workspace_changed_blocks_clone = read_guard.clone();
-                                let post_body = workspace_changed_blocks_clone
-                                    .into_values()
-                                    .collect::<Vec<WorkspaceChangedBlocks>>();
-                                let endpoint = endpoint.clone();
-                                runtime_cloned.spawn(async move {
-                                    let response = client
-                                        .post(endpoint.to_string())
-                                        .json(&post_body)
-                                        .send()
-                                        .await;
-                                    match response {
-                                        Ok(response) => info!(
-                                            "notified hook endpoint, endpoint response status: {}",
-                                            response.status()
-                                        ),
-                                        Err(e) => error!("Failed to send notify: {}", e),
-                                    }
-                                });
-                            }
-                        }
-                        drop(read_guard);
-                        let mut write_guard = workspace_changed_blocks.write().await;
-                        info!("workspace_changed_blocks: {:?}", write_guard);
-                        write_guard.clear();
-                    }
-                });
-                sleep(Duration::from_millis(200));
-            }
-        });
-    }
+    start_handling_observed_blocks(
+        runtime.clone(),
+        workspace_changed_blocks.clone(),
+        hook_endpoint.clone(),
+        client.clone(),
+    );
     let app = files::static_files(sync::sync_handler(api::api_handler(Router::new())))
         .layer(cors)
         .layer(Extension(context.clone()))
